test: fix UserDAO integration tests and add missing jakarta.el dependency

+++ src/main/java/com/example/Service/UserService.java
@@ -7,6 +7,7 @@ import com.example.DAO.UserDAO;
 import com.example.DAO.UserDAOImpl;
 import com.example.DTO.UserDTORequest;
 import com.example.DTO.UserDTOResponse;
+import com.example.configuration.ValidatorUtil;
 import com.example.entity.User;
 import com.example.mapper.UserMapper;
 
@@ -15,21 +16,22 @@ public class UserService {
     private final UserDAO userDao;
     
     private final UserMapper userMapper = new UserMapper();
-
+    
     public UserService(UserDAO userDao) {
         this.userDao = userDao;
     }
 
-    public UserDTOResponse createUser(String name, String email, int age) {
-        if (email == null || !email.contains("@")) {
-            throw new IllegalArgumentException("Некорректный email");
-        }
-        User user = new User(name, email, age);
+    public UserDTOResponse createUser(UserDTORequest dtoRequest) {
+    	ValidatorUtil.validate( dtoRequest );
+    	User user = new User(dtoRequest.name(), dtoRequest.email(), dtoRequest.age());
         userDao.save(user);
         return userMapper.userToDTO(user);
     }
 
     public Optional<UserDTOResponse> findById(Long id) {
+    	if (id == null || id <= 0) {
+    		return Optional.empty();
+        }
         return userDao.findById(id).map( userMapper::userToDTO );
     }
 
@@ -41,8 +43,9 @@ public class UserService {
     }
 
     public void updateUser(Long id, UserDTORequest userDTO) {
+    	ValidatorUtil.validate( userDTO );
     	User user = userDao.findById( id )
-    			.orElseThrow(() -> new IllegalArgumentException("Пользователь с ID " + id + " не найден"));;
+    			.orElseThrow(() -> new IllegalArgumentException("Пользователь с ID " + id + " не найден"));
     	user.setName(userDTO.name());
     	user.setEmail( userDTO.email() );
     	user.setAge(userDTO.age());
@@ -50,6 +53,9 @@ public class UserService {
     }
 
     public void deleteUser(Long id) {
+    	if (id == null || id <= 0) {
+            throw new IllegalArgumentException("Некорректный ID пользователя");
+        }
         userDao.delete(id);
     }
 }
